fix: show tooltip when disabled
